requirements.txt
tensorflow==2.15.0
opencv-python-headless==4.8.1.78
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
scikit-learn
